refactor: split key encoding into lineairdb_keyenc
Move the index key and range encoding functions out of ha_lineairdb.cc into lineairdb_keyenc.{cc,hh} as a pure relocation, keeping them as ha_lineairdb members and placing the shared key-type tag and null-marker constants in the new header, so the handler file holds the storage-engine API rather than encoding internals.
